+++ qflib/pricers/simplepricers.hpp
@@ -15,7 +15,18 @@ BEGIN_NAMESPACE(qf)
 /** The forward price of an asset */
 double fwdPrice(double spot, double timeToExp, double intRate, double divYield);
 
-double quantoFwdPrice(double spot, double timetoExp, double intRate, double divYield, double assetVol, double fxVol, double corr);
+/** Price of a European digital option in the Black-Scholes model*/
+double digitalOptionBS(int payoffType, double spot, double strike, double timeToExp,
+                       double intRate, double divYield, double volatility);
+
+/** Price of a European option in the Black-Scholes model*/
+double europeanOptionBS(int payoffType, double spot, double strike, double timeToExp, 
+                        double intRate, double divYield, double volatility);
+
+/** Price of a Knocout Fwd option */
+double knockoutFwd(double spot, double strike, double KOlevel, double timeToExp, 
+                    double timeToKO, double IntRate, double DivYield, double Vol);
+
 END_NAMESPACE(qf)
 
 #endif // QF_SIMPLEPRICERS_HPP